+++ lib/common/IOStream.h
@@ -41,9 +41,9 @@ public:
 	// Timeout in milliseconds
 	// Read may return 0 -- does not mean end of stream.
 	typedef int64_t pos_type;
-	virtual size_t Read(void *pBuffer, size_t NBytes, int Timeout = IOStream::TimeOutInfinite) = 0;
+	virtual int Read(void *pBuffer, int NBytes, int Timeout = IOStream::TimeOutInfinite) = 0;
 	virtual pos_type BytesLeftToRead();	// may return IOStream::SizeOfStreamUnknown (and will for most stream types)
-	virtual void Write(const void *pBuffer, size_t NBytes) = 0;
+	virtual void Write(const void *pBuffer, int NBytes) = 0;
 	virtual void WriteAllBuffered();
 	virtual pos_type GetPosition() const;
 	virtual void Seek(pos_type Offset, int SeekType);
@@ -55,7 +55,7 @@ public:
 	virtual bool StreamClosed() = 0;
 	
 	// Utility functions
-	bool ReadFullBuffer(void *pBuffer, size_t NBytes, size_t *pNBytesRead, int Timeout = IOStream::TimeOutInfinite);
+	bool ReadFullBuffer(void *pBuffer, int NBytes, int *pNBytesRead, int Timeout = IOStream::TimeOutInfinite);
 	bool CopyStreamTo(IOStream &rCopyTo, int Timeout = IOStream::TimeOutInfinite, int BufferSize = 1024);
 	
 	static int ConvertSeekTypeToOSWhence(int SeekType);
